refactor: uso de ID independiente por entidad

+++ src/main/java/com/institucion/sigea/matricula/repository/MatriculaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MatriculaRepository extends JpaRepository<Matricula, Integer> {
+public interface MatriculaRepository extends JpaRepository<Matricula, Long> {
 
     boolean existsByCodAlumnoAndCodAnioAcademicoAndEstadoTrue(Integer codAlumno, Integer codAnioAcademico);
