+++ src/main/java/com/institucion/sigea/pago/repository/PagoRepository.java
@@ -3,7 +3,7 @@ package com.institucion.sigea.pago.repository;
 import com.institucion.sigea.pago.entity.Pago;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PagoRepository extends JpaRepository<Pago, Integer> {
+public interface PagoRepository extends JpaRepository<Pago, Long> {
 
     boolean existsByCodCuota(Integer codCuota);
 }
