+++ src/main/java/com/institucion/sigea/pago/repository/ReciboRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ReciboRepository extends JpaRepository<Recibo, Integer> {
+public interface ReciboRepository extends JpaRepository<Recibo, Long> {
 
     Optional<Recibo> findByCodPago(Integer codPago);
